linker: garbage collect unused sections but keep multiboot one
This is because libcore, included when Rust uses no_std, has some
dependencies like fmod/fmodf that require some implementation but
are not really used as of now. So instead of implementing them, have
the linker remove unused code referencing them and fail whenever
they are needed so we implement them. At the same time, keep the
multiboot section because even though it is not used by the kernel,
it is used by the boot loader and it would be gc'ed.